MdVTextMesh - monitor character advance and descenders
Handle characters that have an advance < fixedAdvance size so
proportional fonts with smaller advances will center in the character
box, instead of being slammed to the left.

Also handle characters that have ascenders or descenders that overlap
the character box.  This will require some additional effort since some
fonts have a line spacing that does not match the font size.